Bio.Algorithm.Sequence

Description

We emphasize making illegal states unrepresentable. As such, creating something representing a particular kind of genetic string will be done by way of smart constructors. The default constructors will not be exported.

We represent three different kinds of genetic strings (as per Rosalind), DNA, RNA, and Protein.

A DNA is created with mkDNA, an RNA is created with mkRNA, and a Protein is created with mkProtein.

Because we don't export the constructors, pattern matching becomes impossible. To rectify this, we use PatternSynonyms and export patterns for each type. These are PDNA, PRNA, and PProtein, respectively. Internally, we represent sequences as lazy Texts, and this is what you deconstruct one using one of these pattern synonyms.

Documentation

newtype DNA

A DNA sequence. This is constructed using the mkDNA smart constructor below.

Constructors

DNA Text

newtype RNA

An RNA sequence. This is constructed using the mkRNA smart constructor below.

Constructors

RNA Text

class AsDNA a where

There a many different kinds of structures which can be converted into a DNA. We specify these using Prism's.

Methods

_DNA :: Prism' a DNA

isDNAChar :: Char -> Bool

Helper function for helping to validate a valid DNA string.

stringDNA :: Prism' String DNA

lazyTextDNA :: Prism' Text DNA

strictTextDNA :: Prism' Text DNA

lazyBSDNA :: Prism' ByteString DNA

strictBSDNA :: Prism' ByteString DNA