Bio.Algorithm.Types

Description

An experimental approach to working with various bioinformatics types in Haskell. Makes very heavy use of lens.

Making such strong use of lenses, we enable an API that lets you do things like this:

>>> T.pack "CCTTGGAA" ^. lazy . _RawSequence . to (dnaReverseComplement . DNA)
DNA (RawSequence "TTCCAAGG")

Note that you can convert between lazy and strict versions of ByteString and Text by using the appropriate lens combinators.

This module simply re-exports all of Bio.Algorithm.Types.*.

Documentation

module Bio.Algorithm.Types.Protein

module Bio.Algorithm.Types.RawSequence

module Bio.Algorithm.Types.Sequence