Bio.Algorithm.Sequence

Description

Algorithms dealing with bioinformatics sequences. Right now it doesn't do much.

Complements and Reverse Complements

dnaReverseComplement :: DNA -> DNA

Return the reverse complement for some DNA sequence.

>>> dnaReverseComplement (DNA (RawSequence (BL.pack "CCTTGGAA")))
DNA (RawSequence "TTCCAAGG")

>>> T.pack "CCTTGGAA" ^. lazy . _RawSequence . to (dnaReverseComplement . DNA)
DNA (RawSequence "TTCCAAGG")

rnaReverseComplement :: RNA -> RNA

Return the reverse complement for some RNA sequence.

>>> rnaReverseComplement (RNA (RawSequence (BL.pack "GAUGGAACUUGACUACGUAAAUU")))
RNA (RawSequence "AAUUUACGUAGUCAAGUUCCAUC")

>>> T.pack "GAUGGAACUUGACUACGUAAAUU" ^. lazy . _RawSequence . to (rnaReverseComplement . RNA)
RNA (RawSequence "AAUUUACGUAGUCAAGUUCCAUC")

k-mer algorithms

kmers

Arguments

:: AsRawSequence s

=> Int
The length of the k-mers (i.e., the k of the "k-mer")

-> s
The sequence to search in

-> [s]
The resulting list of k-mers

Find and list all of the kmers of the given length in the given string.

>>> kmers 3 (RawSequence $ BL.pack "AGATCGAGTG")
[RawSequence "AGA",RawSequence "GAT",RawSequence "ATC",RawSequence "TCG",RawSequence "CGA",RawSequence "GAG",RawSequence "AGT",RawSequence "GTG"]

>>> kmers 5 (RawSequence $ BL.pack "GTAGAGCTGT")
[RawSequence "GTAGA",RawSequence "TAGAG",RawSequence "AGAGC",RawSequence "GAGCT",RawSequence "AGCTG",RawSequence "GCTGT"]

>>> "gatatat" ^. _RawSequence . to (kmers 4)
[RawSequence "gata",RawSequence "atat",RawSequence "tata",RawSequence "atat"]

>>> "gatatat" ^. _RawSequence . to DNA . to (kmers 4)
[DNA (RawSequence "gata"),DNA (RawSequence "atat"),DNA (RawSequence "tata"),DNA (RawSequence "atat")]

>>> "gatatat" ^. _RawSequence . to RNA . to (kmers 4)
[RNA (RawSequence "gata"),RNA (RawSequence "atat"),RNA (RawSequence "tata"),RNA (RawSequence "atat")]

maxKmers :: (AsRawSequence s, Ord s) => [s] -> [s]

Return only the kmers that appear the highest number of times.

>>> "gatatat" ^. _RawSequence . to (maxKmers . kmers 4)
[RawSequence "atat"]

>>> "gatatat" ^. _RawSequence . to (maxKmers . kmers 4 . DNA)
[DNA (RawSequence "atat")]

>>> "gatatat" ^. _RawSequence . to (maxKmers . kmers 4 . RNA)
[RNA (RawSequence "atat")]

ltClumps

Arguments

:: Int
k. The length of the k-mers (i.e., the k of the "k-mer")

-> Int
l. The interval size to scan

-> Int
t. How many times it must appear

-> RawSequence
The sequence to search in

-> [RawSequence]
The resulting list of k-mers

Find and list all distinct (L, t)-clumps of the given length in the given string.

This is done by scanning the entire genome with a window of size l. In each window, we get all k-mers, see if they exist t times, and if they do, we add them to the result.

>>> "gatcagcataagggtccctgcaatgcatgacaagcctgcagttgttttac" ^. _RawSequence . to (ltClumps 4 25 3)
[RawSequence "tgca"]

Skew

skews :: AsRawSequence s => s -> [Int]

Find skews (difference in number of G and number of C occurances in some sequence) by removing a nucleotide from the right side and recursing over the remaining sequence.

Can be used to solve the Rosalind "Minimum Skew Problem" like this: minSkewIndices s = map (+1) . findIndices (== minimum (skews s)) $ (skews s) although it is too slow to actually be accepted.