Bio.Algorithm.Types

Description

An experimental approach to working with various bioinformatics types in Haskell. Makes very heavy use of lens.

Making such strong use of lenses, we enable an API that lets you do things like this:

>>> T.pack "CCTTGGAA" ^. _RawSequence . to reverseComplement
RawSequence "TTCCAAGG"
it :: RawSequence

Note that you can convert between lazy and strict versions of ByteString and Text by using the appropriate lens combinators.

Documentation

newtype RawSequence

Constructors

RawSequence ByteString

class AsRawSequence p f s where

Minimal complete definition

Nothing

Methods

_RawSequence :: Optic' p f s RawSequence